Add toggle to main window
