+++ Witcher3MapViewer/MainWindow.xaml
@@ -25,6 +25,6 @@
             ItemsSource="{Binding ListOfMaps}"
             SelectedItem="{Binding SelectedMap}"/>
 
-        <local:MarkerToggleControl/>
+        <local:MarkerToggleControl DataContext="{Binding MarkerToggleViewModel}"/>
     </Grid>
 </Window>
